claude: refactor(kaspa): split core module - move bridge logic to new bridge crate
This PR splits the core module into two parts:
- Pure Kaspa functionality stays in libs/kaspa/lib/core
- Bridge/HL-dependent logic moves to new libs/kaspa/lib/bridge crate

The new bridge crate contains:
- message.rs - HL message parsing
- deposit.rs - DepositFXG struct
- withdraw.rs - WithdrawFXG struct
- payload.rs - MessageIDs encoding
- confirmation.rs - ConfirmationFXG struct
- util.rs - Address<->H256 conversion
- user/ - User deposit/payload utilities

After this change:
- libs/kaspa/lib/core has ZERO hyperlane dependencies
- libs/kaspa/lib/bridge depends on both core (pure Kaspa) and HL
- relayer/validator/tooling updated to use bridge crate
- dymension-kaspa re-exports bridge as kas_bridge

This completes Phase 2 of dependency inversion.

Refs: #140
