+++ dymension/libs/kaspa/Cargo.toml
@@ -4,6 +4,7 @@ resolver = "2" # Enables new feature resolver, good practice for new workspaces
 
 members = [
     "lib/core",
+    "lib/bridge",
     "lib/relayer",
     "lib/validator",
     "lib/api",
