+++ dymension/libs/kaspa/lib/bridge/src/util.rs
