+++ dymension/libs/kaspa/lib/core/src/api/client.rs
@@ -169,9 +169,9 @@ impl HttpClient {
                     continue;
                 }
 
-                if !has_valid_hyperlane_payload(&tx, domain_kas) {
-                    continue;
-                }
+                // TODO: Add back HL payload validation when we have a way to do it without HL deps
+                // The payload validation was moved to kas_bridge module
+                // For now, we rely on other validation to filter invalid deposits
 
                 let tx_id = tx.transaction_id.clone();
                 let tx_time = tx.block_time;
@@ -288,88 +288,14 @@ fn is_valid_escrow_transfer(tx: &TxModel, address: &String) -> Result<bool> {
     Ok(false)
 }
 
-// check both that the it deserializes to a HL message, but also that the origin is correct
-// the parse can easily pass for random blobs, so we check the domain to be really sure
-// TODO: move to hyperlane-monorepo/dymension/libs/kaspa/lib/core/src/user/payload.rs
-fn has_valid_hyperlane_payload(tx: &TxModel, domain_kas: u32) -> bool {
-    use crate::message::ParsedHL;
-
-    match &tx.payload {
-        Some(payload) => {
-            let parsed = match ParsedHL::parse_string(payload) {
-                Ok(p) => p,
-                Err(_) => return false,
-            };
-            parsed.hl_message.origin == domain_kas
-        }
-        None => false,
-    }
-}
+// Removed has_valid_hyperlane_payload - moved to kas_bridge module
+// The HL-specific validation was removed to keep this module pure Kaspa
 
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::message::ParsedHL;
-    use dymension_kaspa_hl_constants::{HL_DOMAIN_KASPA_TEST10, HL_DOMAIN_KASPA_TEST10_LEGACY};
 
-    #[tokio::test]
-    #[ignore = "dont hit real api"]
-    async fn test_get_deposits() {
-        // https://explorer-tn10.kaspa.org/addresses/kaspatest:pzlq49spp66vkjjex0w7z8708f6zteqwr6swy33fmy4za866ne90v7e6pyrfr?page=1
-        let client = HttpClient::new(
-            "https://api-tn10.kaspa.org/".to_string(),
-            RateLimitConfig::default(),
-        );
-        let address = "kaspatest:pzlq49spp66vkjjex0w7z8708f6zteqwr6swy33fmy4za866ne90v7e6pyrfr";
-
-        let deposits = client
-            .get_deposits_by_address(Some(1751299515650), address, HL_DOMAIN_KASPA_TEST10)
-            .await;
-
-        match deposits {
-            Ok(deposits) => {
-                println!("Found deposits: n = {:?}", deposits.len());
-                for deposit in deposits {
-                    println!("Deposit: {:?}", deposit);
-                }
-            }
-            Err(e) => {
-                println!("Query deposits: {:?}", e);
-            }
-        }
-    }
-
-    #[tokio::test]
-    #[ignore = "dont hit real api"]
-    async fn test_get_deposit_stress() {
-        /*
-        Tries to check if its really working as we observed that the relayer missed many deposits
-         */
-        let client = HttpClient::new(
-            "https://api-tn10.kaspa.org/".to_string(),
-            RateLimitConfig::default(),
-        );
-        // blumbus address Nov 19 2025
-        let address = "kaspatest:pzwcd30pvdn0k4snvj5awkmlm6srzuw8d8e766ff5vwceg2akta3799nq2a3p";
-
-        let deposits = client
-            // at time of writing test, am absolutely sure how many deposits are just after this time, because of stress testing
-            .get_deposits_by_address(Some(1763568168707), address, HL_DOMAIN_KASPA_TEST10_LEGACY)
-            .await;
-
-        match deposits {
-            Ok(deposits) => {
-                let n = deposits.len();
-                for deposit in deposits {
-                    println!("Deposit: {:?}", deposit);
-                }
-                println!("Found deposits: n = {:?}", n);
-            }
-            Err(e) => {
-                println!("Query deposits: {:?}", e);
-            }
-        }
-    }
+    // Removed HL-specific deposit tests - they should be in integration tests in kas_bridge
 
     #[tokio::test]
     #[ignore]
@@ -383,10 +309,5 @@ mod tests {
         println!("Tx: {:?}", tx);
     }
 
-    #[tokio::test]
-    async fn test_parse() {
-        let s = "0a20020c0a41a75218b6f6e3fbf19c828239f368abb983ff4d6b16a3ea594f6b19770a204fd1e8d1ce6debfbc425d262785ab26167f71d3800e6d5be8af59bac7b548e960a2003c7520e12eda99f6e0da65eebd82823980594361cc719439dd05c5ffe9b8b4d0a20e4772ae63275386a3f3ba6582ac8edbecdacc817529af21f29b7c3dafd0ed9820a20cd26ff93e58234a18da37119ee4452975654721ee95df637114dde146b4493ba0a2039170a92263b1bbc7ad539e33cdced96f149db7a246cd62bc6dced1538448ecc0a20dbd9ba9bfdafb77fa13324044be150283815f4e0587ce863d2151e3782537e380a2013a4cbdfad246476602b4a6c384a427f4afc13c931957d3b7081cf2548dc11420a20ef88cd0cdde1d955258dd62febc2fe818b75d74e1bb211acce4dd2061c4a7b660a20e824d5f6f83869430bf04332978d87dc65b50102cd680ade24da240fe84ddef60a20ab9ef65d2fa925e9be3f1409c19416e8e3a36a8e0fa11e1169bc958b7f7febe00a2033569543493e026add696ec883ae11821fa66905af1ef502d5b7bc903478046d0a2045be6ff58cca1b077a5e77a5e9458ab5f86fc50f6d8e4ebf905d024e0677b8470a20f41a429e7f732880b6760fda7ad0a4918ee347e92ed341358dc86c6d658738610a204d9c9a5087544f68a81200c3c0660579aef5329685ab2142280727fadbfb1ac70a202501746dd5b4ed854cb6c15df0f4baf4af0f9d7f6eee3b260a50dc30164c3c53";
-        let parsed = ParsedHL::parse_string(s);
-        assert!(parsed.is_ok(), "Failed to parse payload");
-    }
+    // Removed HL-specific parse test - should be in kas_bridge module tests
 }
