+++ dymension/libs/kaspa/lib/core/src/lib.rs
@@ -1,17 +1,10 @@
 pub mod api;
 pub mod balance;
-pub mod confirmation;
 pub mod consts;
-pub mod deposit;
 pub mod env;
 pub mod escrow;
 pub mod finality;
-pub mod message;
-pub mod payload;
 pub mod pskt;
-pub mod user;
-pub mod util;
 pub mod wallet;
-pub mod withdraw;
 
 pub use secp256k1::Keypair as KaspaSecpKeypair;
