+++ dymension/libs/kaspa/lib/relayer/src/confirm/confirmation.rs
@@ -1,5 +1,5 @@
 use corelib::api::client::HttpClient;
-use corelib::{confirmation::ConfirmationFXG, payload::MessageID};
+use bridge::{confirmation::ConfirmationFXG, payload::MessageID};
 use eyre::Result;
 use hex;
 use kaspa_consensus_core::tx::TransactionOutpoint;
@@ -131,7 +131,7 @@ pub async fn recursive_trace_transactions(
             .clone()
             .ok_or_else(|| eyre::eyre!("No payload found in transaction"))?;
 
-        let message_ids = corelib::payload::MessageIDs::from_tx_payload(&payload)?;
+        let message_ids = bridge::payload::MessageIDs::from_tx_payload(&payload)?;
 
         processed_withdrawals.extend(message_ids.0);
         outpoint_sequence.push(out_curr);
