+++ dymension/libs/kaspa/lib/relayer/src/withdraw/tooling.rs
@@ -1,5 +1,5 @@
 use corelib::escrow::*;
-use corelib::util::input_sighash_type;
+use bridge::util::input_sighash_type;
 use eyre::Result;
 use kaspa_addresses::Address;
 use kaspa_consensus_core::tx::{TransactionOutpoint, UtxoEntry};
