+++ dymension/libs/kaspa/tooling/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2021"
 [dependencies]
 kaspa-core = { workspace = true }
 corelib = { path = "../lib/core", package = "core" }
+bridge = { path = "../lib/bridge" }
 hardcode = { path = "../lib/hardcode" }
 validator = { path = "../lib/validator" }
 relayer = { path = "../lib/relayer" }
