+++ rust/main/chains/dymension-kaspa/Cargo.toml
@@ -52,6 +52,7 @@ axum = { workspace = true }
 dym-kas-validator = { path = "../../../../dymension/libs/kaspa/lib/validator", package="validator"}
 dym-kas-relayer = { path = "../../../../dymension/libs/kaspa/lib/relayer" , package="relayer"}
 dym-kas-core = { path = "../../../../dymension/libs/kaspa/lib/core" , package="core"}
+dym-kas-bridge = { path = "../../../../dymension/libs/kaspa/lib/bridge" , package="bridge"}
 dym-kas-api = { path = "../../../../dymension/libs/kaspa/lib/api" , package="openapi"}
 dym-kas-kms = { path = "../../../../dymension/libs/kaspa/lib/kms", package="dym-kas-kms"}
 dymension-kaspa-hl-constants = { path = "../dymension-kaspa-hl-constants" }
